Greatly improved popup behaviour
Popup will be automatically triggered when a point is detected, but only once. It can be closed both by selecting an option, or pressing the key ESC. 
The popup toggle is set back on as soon as the text changes.
[Scenario - ContinuumAutocompletePopup = CAP]
myVar.
(CAP On)
ESC
(Cap Off)
Write S
Delete S
(CAP On)